requirements.txt
fastapi
uvicorn[standard]
asyncpg
humps
httpx
python-multipart
pyyaml
Jinja2
beautifulsoup4
pydantic-settings
ollama
